- use turboframes to lazy load transfers with > 0 depth   - revert to using erb for group show page   - use activity_feed route and controller (from demo), refactor later   - break up the show view rb file into smaller components (rendered from erb) - move summarise_for to group model
