+++ app/controllers/activity_feed_controller.rb
@@ -1,0 +1,5 @@
+class ActivityFeedController < ApplicationController
+  def show
+    @group = Group.find(params[:id])
+  end
+end
