+++ app/controllers/groups_controller.rb
@@ -12,7 +12,6 @@ class GroupsController < ApplicationController
 
   # GET /groups/1 or /groups/1.json
   def show
-    render Groups::ShowView.new(group: @group)
   end
 
   # GET /groups/new
