+++ app/views/activity_feed/show.html.erb
@@ -1,0 +1,12 @@
+<%= turbo_frame_tag :feed do %>
+  <div>
+    <%= render TransfersTableComponent.new(
+        entity: @group,
+        transfers: @group.consolidated_transfers(depth: 6),
+        heading: "Connected to groups and people that are connected to #{@group.name} to a depth of #{6} degrees of separation",
+        summarise_for: Group.summarise_for(@group),
+        remove_zero_degrees: true
+      )
+    %>
+  </div>
+<% end %>
