+++ app/views/components/search_results.rb
